// app/generated/model/client/bench/protocol/messages.ts
export interface C2S_MailboxParity extends IRequest {
  rpcId?: number;
  callCount: number;
  delayMs: number;
}

export const C2S_MailboxParityCodec = {
  decode(payload: Uint8Array): C2S_MailboxParity {
    const reader = new BinaryReader(payload);
    const value: C2S_MailboxParity = {
      callCount: 0,
      delayMs: 0,
    };
    while (!reader.eof()) {
      const tag = reader.tag();
      if (tag.fieldNo === 90 && tag.wireType === 0) {
        value.rpcId = reader.uint32();
      }
      else if (tag.fieldNo === 1 && tag.wireType === 0) {
        value.callCount = reader.uint32();
      }
      else if (tag.fieldNo === 2 && tag.wireType === 0) {
        value.delayMs = reader.uint32();
      }
      else {
        reader.skip(tag.wireType);
      }
    }
    return value;
  },

  encode(value: C2S_MailboxParity): Uint8Array {
    const writer = new BinaryWriter();
    if (value.rpcId !== undefined) writer.uint32(90, value.rpcId);
    writer.uint32(1, value.callCount);
    writer.uint32(2, value.delayMs);
    return writer.finish();
  },
};

export interface S2C_MailboxParity extends IResponse {
  message?: string;
  error?: number;
  rpcId?: number;
  elapsedMs: number;
  maxServerConcurrency: number;
}

export const S2C_MailboxParityCodec = {
  decode(payload: Uint8Array): S2C_MailboxParity {
    const reader = new BinaryReader(payload);
    const value: S2C_MailboxParity = {
      elapsedMs: 0,
      maxServerConcurrency: 0,
    };
    while (!reader.eof()) {
      const tag = reader.tag();
      if (tag.fieldNo === 92 && tag.wireType === 2) {
        value.message = reader.string();
      }
      else if (tag.fieldNo === 91 && tag.wireType === 0) {
        value.error = reader.uint32();
      }
      else if (tag.fieldNo === 90 && tag.wireType === 0) {
        value.rpcId = reader.uint32();
      }
      else if (tag.fieldNo === 1 && tag.wireType === 0) {
        value.elapsedMs = reader.uint32();
      }
      else if (tag.fieldNo === 2 && tag.wireType === 0) {
        value.maxServerConcurrency = reader.uint32();
      }
      else {
        reader.skip(tag.wireType);
      }
    }
    return value;
  },

  encode(value: S2C_MailboxParity): Uint8Array {
    const writer = new BinaryWriter();
    if (value.message !== undefined) writer.string(92, value.message);
    if (value.error !== undefined) writer.uint32(91, value.error);
    if (value.rpcId !== undefined) writer.uint32(90, value.rpcId);
    writer.uint32(1, value.elapsedMs);
    writer.uint32(2, value.maxServerConcurrency);
    return writer.finish();
  },
};