// app/generated/model/client/bench/protocol/rpcs.ts
// Generated by tools/codegen_proto.mjs. Do not edit by hand.
import { defineRpc } from "../../../../../core/protocol/rpc";
import {
  C2S_MailboxParity,
  C2S_MailboxParityCodec,
  C2S_RuntimePing,
  C2S_RuntimePingCodec,
  S2C_MailboxParity,
  S2C_MailboxParityCodec,
  S2C_RuntimePing,
  S2C_RuntimePingCodec,
} from "./messages";
import { MsgCode } from "./msgcodes";

export const BenchProtocol = {
  RuntimePing: defineRpc<C2S_RuntimePing, S2C_RuntimePing>({
    name: "Bench.RuntimePing",
    requestCode: MsgCode.C2S_RuntimePing,
    responseCode: MsgCode.S2C_RuntimePing,
    requestCodec: C2S_RuntimePingCodec,
    responseCodec: S2C_RuntimePingCodec,
  }),
};

export const MailboxParityProtocol = {
  MailboxParity: defineRpc<C2S_MailboxParity, S2C_MailboxParity>({
    name: "MailboxParity.MailboxParity",
    requestCode: MsgCode.C2S_MailboxParity,
    responseCode: MsgCode.S2C_MailboxParity,
    requestCodec: C2S_MailboxParityCodec,
    responseCodec: S2C_MailboxParityCodec,
  }),
};

export const AllRpcDescriptors = [
  BenchProtocol.RuntimePing,
  MailboxParityProtocol.MailboxParity,
] as const;